RPi_data/requirements.txt
appdirs==1.4.0
cffi==1.9.1
click==6.7
cryptography==1.7.2
enum-compat==0.0.2
eventlet==0.20.1
Flask==0.12
Flask-SocketIO==2.8.2
greenlet==0.4.12
idna==2.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
paho-mqtt==1.2
pyasn1==0.2.2
pycparser==2.17
pymongo==3.4.0
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-engineio==1.1.2
python-socketio==1.6.3
requests==2.13.0
six==1.10.0
virtualenv==15.1.0
Werkzeug==0.11.15